% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Advice}
\alias{Advice}
\title{Google Service Management API Objects 
Google Service Management allows service producers to publish their services on Google Cloud Platform so that they can be discovered and used by service consumers.}
\usage{
Advice(description = NULL)
}
\arguments{
\item{description}{Useful description for why this advice was applied and what actions should}
}
\value{
Advice object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:12:36
filename: /Users/mark/dev/R/autoGoogleAPI/googleservicemanagementv1.auto/R/servicemanagement_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Advice Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Generated advice about this change, used for providing moreinformation about how a change will affect the existing service.
}
