% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Api}
\alias{Api}
\title{Api Object}
\usage{
Api(name = NULL, sourceContext = NULL, syntax = NULL, version = NULL,
  mixins = NULL, options = NULL, methods = NULL)
}
\arguments{
\item{name}{The fully qualified name of this api, including package name}

\item{sourceContext}{Source context for the protocol buffer service represented by this}

\item{syntax}{The source syntax of the service}

\item{version}{A version string for this api}

\item{mixins}{Included APIs}

\item{options}{Any metadata attached to the API}

\item{methods}{The methods of this api, in unspecified order}
}
\value{
Api object
}
\description{
Api Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Api is a light-weight descriptor for a protocol buffer service.
}
