% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{AuthorizationConfig}
\alias{AuthorizationConfig}
\title{AuthorizationConfig Object}
\usage{
AuthorizationConfig(provider = NULL)
}
\arguments{
\item{provider}{The name of the authorization provider, such as}
}
\value{
AuthorizationConfig object
}
\description{
AuthorizationConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration of authorization.This section determines the authorization provider, if unspecified, then noauthorization check will be done.Example:    experimental:      authorization:        provider: firebaserules.googleapis.com
}
