% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{AuthProvider}
\alias{AuthProvider}
\title{AuthProvider Object}
\usage{
AuthProvider(jwksUri = NULL, audiences = NULL, id = NULL, issuer = NULL)
}
\arguments{
\item{jwksUri}{URL of the provider's public key set to validate signature of the JWT}

\item{audiences}{The list of JWT}

\item{id}{The unique identifier of the auth provider}

\item{issuer}{Identifies the principal that issued the JWT}
}
\value{
AuthProvider object
}
\description{
AuthProvider Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for an anthentication provider, including support for[JSON Web Token (JWT)](https://tools.ietf.org/html/draft-ietf-oauth-json-web-token-32).
}
