% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{AuthRequirement}
\alias{AuthRequirement}
\title{AuthRequirement Object}
\usage{
AuthRequirement(audiences = NULL, providerId = NULL)
}
\arguments{
\item{audiences}{NOTE: This will be deprecated soon, once AuthProvider}

\item{providerId}{id from authentication provider}
}
\value{
AuthRequirement object
}
\description{
AuthRequirement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User-defined authentication requirements, including support for[JSON Web Token (JWT)](https://tools.ietf.org/html/draft-ietf-oauth-json-web-token-32).
}
