% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{ChangeReport}
\alias{ChangeReport}
\title{ChangeReport Object}
\usage{
ChangeReport(configChanges = NULL)
}
\arguments{
\item{configChanges}{List of changes between two service configurations}
}
\value{
ChangeReport object
}
\description{
ChangeReport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Change report associated with a particular service configuration.It contains a list of ConfigChanges based on the comparison betweentwo service configurations.
}
