% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Condition}
\alias{Condition}
\title{Condition Object}
\usage{
Condition(sys = NULL, value = NULL, iam = NULL, values = NULL,
  op = NULL, svc = NULL)
}
\arguments{
\item{sys}{Trusted attributes supplied by any service that owns resources and uses}

\item{value}{DEPRECATED}

\item{iam}{Trusted attributes supplied by the IAM system}

\item{values}{The objects of the condition}

\item{op}{An operator to apply the subject with}

\item{svc}{Trusted attributes discharged by the service}
}
\value{
Condition object
}
\description{
Condition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A condition to be met.
}
