% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{ConfigChange}
\alias{ConfigChange}
\title{ConfigChange Object}
\usage{
ConfigChange(newValue = NULL, changeType = NULL, element = NULL,
  oldValue = NULL, advices = NULL)
}
\arguments{
\item{newValue}{Value of the changed object in the new Service configuration,}

\item{changeType}{The type for this change, either ADDED, REMOVED, or MODIFIED}

\item{element}{Object hierarchy path to the change, with levels separated by a '}

\item{oldValue}{Value of the changed object in the old Service configuration,}

\item{advices}{Collection of advice provided for this change, useful for determining the}
}
\value{
ConfigChange object
}
\description{
ConfigChange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Output generated from semantically comparing two versions of a serviceconfiguration.Includes detailed information about a field that have changed withapplicable advice about potential consequences for the change, such asbackwards-incompatibility.
}
