% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{ConfigFile}
\alias{ConfigFile}
\title{ConfigFile Object}
\usage{
ConfigFile(fileType = NULL, fileContents = NULL, filePath = NULL)
}
\arguments{
\item{fileType}{The type of configuration file this represents}

\item{fileContents}{The bytes that constitute the file}

\item{filePath}{The file name of the configuration file (full or relative path)}
}
\value{
ConfigFile object
}
\description{
ConfigFile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Generic specification of a source configuration file
}
