% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{ConfigRef}
\alias{ConfigRef}
\title{ConfigRef Object}
\usage{
ConfigRef(name = NULL)
}
\arguments{
\item{name}{Resource name of a service config}
}
\value{
ConfigRef object
}
\description{
ConfigRef Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a service configuration with its name and id.
}
