% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{ConfigSource}
\alias{ConfigSource}
\title{ConfigSource Object}
\usage{
ConfigSource(files = NULL, id = NULL)
}
\arguments{
\item{files}{Set of source configuration files that are used to generate a service}

\item{id}{A unique ID for a specific instance of this message, typically assigned}
}
\value{
ConfigSource object
}
\description{
ConfigSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a source file which is used to generate the service configurationdefined by `google.api.Service`.
}
