% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Context}
\alias{Context}
\title{Context Object}
\usage{
Context(rules = NULL)
}
\arguments{
\item{rules}{A list of RPC context rules that apply to individual API methods}
}
\value{
Context object
}
\description{
Context Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Context` defines which contexts an API requests.Example:    context:      rules:      - selector: '*'        requested:        - google.rpc.context.ProjectContext        - google.rpc.context.OriginContextThe above specifies that all methods in the API request`google.rpc.context.ProjectContext` and`google.rpc.context.OriginContext`.Available context types are defined in package`google.rpc.context`.
}
