% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{ContextRule}
\alias{ContextRule}
\title{ContextRule Object}
\usage{
ContextRule(requested = NULL, selector = NULL, provided = NULL)
}
\arguments{
\item{requested}{A list of full type names of requested contexts}

\item{selector}{Selects the methods to which this rule applies}

\item{provided}{A list of full type names of provided contexts}
}
\value{
ContextRule object
}
\description{
ContextRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A context rule provides information about the context for an individual APIelement.
}
