% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Control}
\alias{Control}
\title{Control Object}
\usage{
Control(environment = NULL)
}
\arguments{
\item{environment}{The service control environment to use}
}
\value{
Control object
}
\description{
Control Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Selects and configures the service controller used by the service.  Theservice controller handles features like abuse, quota, billing, logging,monitoring, etc.
}
