% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{CustomErrorRule}
\alias{CustomErrorRule}
\title{CustomErrorRule Object}
\usage{
CustomErrorRule(isErrorType = NULL, selector = NULL)
}
\arguments{
\item{isErrorType}{Mark this message as possible payload in error response}

\item{selector}{Selects messages to which this rule applies}
}
\value{
CustomErrorRule object
}
\description{
CustomErrorRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A custom error rule.
}
