% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{DeleteServiceStrategy}
\alias{DeleteServiceStrategy}
\title{DeleteServiceStrategy Object}
\usage{
DeleteServiceStrategy()
}
\value{
DeleteServiceStrategy object
}
\description{
DeleteServiceStrategy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Strategy used to delete a service. This strategy is a placeholder onlyused by the system generated rollout to delete a service.
}
