% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Diagnostic}
\alias{Diagnostic}
\title{Diagnostic Object}
\usage{
Diagnostic(message = NULL, location = NULL)
}
\arguments{
\item{message}{Message describing the error or warning}

\item{location}{File name and line number of the error or warning}
}
\value{
Diagnostic object
}
\description{
Diagnostic Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a diagnostic message (error or warning)
}
