% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{DisableServiceRequest}
\alias{DisableServiceRequest}
\title{DisableServiceRequest Object}
\usage{
DisableServiceRequest(consumerId = NULL)
}
\arguments{
\item{consumerId}{The identity of consumer resource which service disablement will be}
}
\value{
DisableServiceRequest object
}
\description{
DisableServiceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for DisableService method.
}
\seealso{
Other DisableServiceRequest functions: \code{\link{services.disable}}
}
