% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{DocumentationRule}
\alias{DocumentationRule}
\title{DocumentationRule Object}
\usage{
DocumentationRule(deprecationDescription = NULL, selector = NULL,
  description = NULL)
}
\arguments{
\item{deprecationDescription}{Deprecation description of the selected element(s)}

\item{selector}{The selector is a comma-separated list of patterns}

\item{description}{Description of the selected API(s)}
}
\value{
DocumentationRule object
}
\description{
DocumentationRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A documentation rule provides information about individual API elements.
}
