% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{EnumValue}
\alias{EnumValue}
\title{EnumValue Object}
\usage{
EnumValue(number = NULL, name = NULL, options = NULL)
}
\arguments{
\item{number}{Enum value number}

\item{name}{Enum value name}

\item{options}{Protocol buffer options}
}
\value{
EnumValue object
}
\description{
EnumValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Enum value definition.
}
