% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Field}
\alias{Field}
\title{Field Object}
\usage{
Field(name = NULL, typeUrl = NULL, number = NULL, jsonName = NULL,
  options = NULL, oneofIndex = NULL, cardinality = NULL, packed = NULL,
  defaultValue = NULL)
}
\arguments{
\item{name}{The field name}

\item{typeUrl}{The field type URL, without the scheme, for message or enumeration}

\item{number}{The field number}

\item{jsonName}{The field JSON name}

\item{options}{The protocol buffer options}

\item{oneofIndex}{The index of the field type in `Type}

\item{cardinality}{The field cardinality}

\item{packed}{Whether to use alternative packed wire representation}

\item{defaultValue}{The string value of the default value of this field}
}
\value{
Field object
}
\description{
Field Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single field of a message type.
}
