% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{GenerateConfigReportRequest.oldConfig}
\alias{GenerateConfigReportRequest.oldConfig}
\title{GenerateConfigReportRequest.oldConfig Object}
\usage{
GenerateConfigReportRequest.oldConfig()
}
\value{
GenerateConfigReportRequest.oldConfig object
}
\description{
GenerateConfigReportRequest.oldConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Service configuration against which the comparison will be done.For this version of API, the supported types aregoogle.api.servicemanagement.v1.ConfigRef,google.api.servicemanagement.v1.ConfigSource,and google.api.Service
}
\seealso{
Other GenerateConfigReportRequest functions: \code{\link{GenerateConfigReportRequest.newConfig}},
  \code{\link{GenerateConfigReportRequest}},
  \code{\link{services.generateConfigReport}}
}
