% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{GenerateConfigReportRequest}
\alias{GenerateConfigReportRequest}
\title{GenerateConfigReportRequest Object}
\usage{
GenerateConfigReportRequest(GenerateConfigReportRequest.oldConfig = NULL,
  GenerateConfigReportRequest.newConfig = NULL, oldConfig = NULL,
  newConfig = NULL)
}
\arguments{
\item{GenerateConfigReportRequest.oldConfig}{The \link{GenerateConfigReportRequest.oldConfig} object or list of objects}

\item{GenerateConfigReportRequest.newConfig}{The \link{GenerateConfigReportRequest.newConfig} object or list of objects}

\item{oldConfig}{Service configuration against which the comparison will be done}

\item{newConfig}{Service configuration for which we want to generate the report}
}
\value{
GenerateConfigReportRequest object
}
\description{
GenerateConfigReportRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for GenerateConfigReport method.
}
\seealso{
Other GenerateConfigReportRequest functions: \code{\link{GenerateConfigReportRequest.newConfig}},
  \code{\link{GenerateConfigReportRequest.oldConfig}},
  \code{\link{services.generateConfigReport}}
}
