% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{GenerateConfigReportResponse}
\alias{GenerateConfigReportResponse}
\title{GenerateConfigReportResponse Object}
\usage{
GenerateConfigReportResponse(id = NULL, diagnostics = NULL,
  serviceName = NULL, changeReports = NULL)
}
\arguments{
\item{id}{ID of the service configuration this report belongs to}

\item{diagnostics}{Errors / Linter warnings associated with the service definition this}

\item{serviceName}{Name of the service this report belongs to}

\item{changeReports}{list of ChangeReport, each corresponding to comparison between two}
}
\value{
GenerateConfigReportResponse object
}
\description{
GenerateConfigReportResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for GenerateConfigReport method.
}
