% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{ListServiceConfigsResponse}
\alias{ListServiceConfigsResponse}
\title{ListServiceConfigsResponse Object}
\usage{
ListServiceConfigsResponse(serviceConfigs = NULL, nextPageToken = NULL)
}
\arguments{
\item{serviceConfigs}{The list of service configuration resources}

\item{nextPageToken}{The token of the next page of results}
}
\value{
ListServiceConfigsResponse object
}
\description{
ListServiceConfigsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ListServiceConfigs method.
}
