% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{ListServiceRolloutsResponse}
\alias{ListServiceRolloutsResponse}
\title{ListServiceRolloutsResponse Object}
\usage{
ListServiceRolloutsResponse(nextPageToken = NULL, rollouts = NULL)
}
\arguments{
\item{nextPageToken}{The token of the next page of results}

\item{rollouts}{The list of rollout resources}
}
\value{
ListServiceRolloutsResponse object
}
\description{
ListServiceRolloutsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for ListServiceRollouts method.
}
