% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{ListServicesResponse}
\alias{ListServicesResponse}
\title{ListServicesResponse Object}
\usage{
ListServicesResponse(services = NULL, nextPageToken = NULL)
}
\arguments{
\item{services}{The returned services will only have the name field set}

\item{nextPageToken}{Token that can be passed to `ListServices` to resume a paginated query}
}
\value{
ListServicesResponse object
}
\description{
ListServicesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for `ListServices` method.
}
