% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{ManagedService}
\alias{ManagedService}
\title{ManagedService Object}
\usage{
ManagedService(serviceName = NULL, producerProjectId = NULL)
}
\arguments{
\item{serviceName}{The name of the service}

\item{producerProjectId}{ID of the project that produces and owns this service}
}
\value{
ManagedService object
}
\description{
ManagedService Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The full representation of a Service that is managed byGoogle Service Management.
}
\seealso{
Other ManagedService functions: \code{\link{services.create}}
}
