% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{MediaUpload}
\alias{MediaUpload}
\title{MediaUpload Object}
\usage{
MediaUpload(uploadService = NULL, enabled = NULL)
}
\arguments{
\item{uploadService}{DO NOT USE THIS FIELD UNTIL THIS WARNING IS REMOVED}

\item{enabled}{Whether upload is enabled}
}
\value{
MediaUpload object
}
\description{
MediaUpload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Use this only for Scotty Requests. Do not use this for media support usingBytestream, add instead [][google.bytestream.RestByteStream] as an API toyour configuration for Bytestream methods.
}
