% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Monitoring}
\alias{Monitoring}
\title{Monitoring Object}
\usage{
Monitoring(consumerDestinations = NULL, producerDestinations = NULL)
}
\arguments{
\item{consumerDestinations}{Monitoring configurations for sending metrics to the consumer project}

\item{producerDestinations}{Monitoring configurations for sending metrics to the producer project}
}
\value{
Monitoring object
}
\description{
Monitoring Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Monitoring configuration of the service.The example below shows how to configure monitored resources and metricsfor monitoring. In the example, a monitored resource and two metrics aredefined. The `library.googleapis.com/book/returned_count` metric is sentto both producer and consumer projects, whereas the`library.googleapis.com/book/overdue_count` metric is only sent to theconsumer project.    monitored_resources:    - type: library.googleapis.com/branch      labels:      - key: /city        description: The city where the library branch is located in.      - key: /name        description: The name of the branch.    metrics:    - name: library.googleapis.com/book/returned_count      metric_kind: DELTA      value_type: INT64      labels:      - key: /customer_id    - name: library.googleapis.com/book/overdue_count      metric_kind: GAUGE      value_type: INT64      labels:      - key: /customer_id    monitoring:      producer_destinations:      - monitored_resource: library.googleapis.com/branch        metrics:        - library.googleapis.com/book/returned_count      consumer_destinations:      - monitored_resource: library.googleapis.com/branch        metrics:        - library.googleapis.com/book/returned_count        - library.googleapis.com/book/overdue_count
}
