% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{MonitoringDestination}
\alias{MonitoringDestination}
\title{MonitoringDestination Object}
\usage{
MonitoringDestination(monitoredResource = NULL, metrics = NULL)
}
\arguments{
\item{monitoredResource}{The monitored resource type}

\item{metrics}{Names of the metrics to report to this monitoring destination}
}
\value{
MonitoringDestination object
}
\description{
MonitoringDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration of a specific monitoring destination (the producer projector the consumer project).
}
