% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{OperationMetadata}
\alias{OperationMetadata}
\title{OperationMetadata Object}
\usage{
OperationMetadata(startTime = NULL, resourceNames = NULL, steps = NULL,
  progressPercentage = NULL)
}
\arguments{
\item{startTime}{The start time of the operation}

\item{resourceNames}{The full name of the resources that this operation is directly}

\item{steps}{Detailed status information for each step}

\item{progressPercentage}{Percentage of completion of this operation, ranging from 0 to 100}
}
\value{
OperationMetadata object
}
\description{
OperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata associated with a long running operation resource.
}
