% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Option}
\alias{Option}
\title{Option Object}
\usage{
Option(Option.value = NULL, value = NULL, name = NULL)
}
\arguments{
\item{Option.value}{The \link{Option.value} object or list of objects}

\item{value}{The option's value packed in an Any message}

\item{name}{The option's name}
}
\value{
Option object
}
\description{
Option Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A protocol buffer option, which can be attached to a message, field,enumeration, etc.
}
\seealso{
Other Option functions: \code{\link{Option.value}}
}
