% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Rollout}
\alias{Rollout}
\title{Rollout Object}
\usage{
Rollout(rolloutId = NULL, deleteServiceStrategy = NULL, createTime = NULL,
  status = NULL, serviceName = NULL, trafficPercentStrategy = NULL,
  createdBy = NULL)
}
\arguments{
\item{rolloutId}{Optional unique identifier of this Rollout}

\item{deleteServiceStrategy}{The strategy associated with a rollout to delete a `ManagedService`}

\item{createTime}{Creation time of the rollout}

\item{status}{The status of this rollout}

\item{serviceName}{The name of the service associated with this Rollout}

\item{trafficPercentStrategy}{Google Service Control selects service configurations based on}

\item{createdBy}{The user who created the Rollout}
}
\value{
Rollout object
}
\description{
Rollout Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A rollout resource that defines how service configuration versions are pushedto control plane systems. Typically, you create a new version of theservice config, and then create a Rollout to push the service config.
}
