% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_functions.R
\docType{package}
\name{servicemanagement_googleAuthR}
\alias{servicemanagement_googleAuthR}
\alias{servicemanagement_googleAuthR-package}
\title{Google Service Management API
Google Service Management allows service producers to publish their services on Google Cloud Platform so that they can be discovered and used by service consumers.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:12:36
filename: /Users/mark/dev/R/autoGoogleAPI/googleservicemanagementv1.auto/R/servicemanagement_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/service.management
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/service.management.readonly
\item https://www.googleapis.com/auth/cloud-platform
}
}
