% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_functions.R
\name{services.delete}
\alias{services.delete}
\title{Deletes a managed service. This method will change the service to the`Soft-Delete` state for 30 days. Within this period, service producers maycall UndeleteService to restore the service.After 30 days, the service will be permanently deleted.Operation<response: google.protobuf.Empty>}
\usage{
services.delete(serviceName)
}
\arguments{
\item{serviceName}{The name of the service}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/service.management
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/service.management)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-management/}{Google Documentation}
}
