% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_functions.R
\name{services.disable}
\alias{services.disable}
\title{Disable a managed service for a project.Operation<response: DisableServiceResponse>}
\usage{
services.disable(DisableServiceRequest, serviceName)
}
\arguments{
\item{DisableServiceRequest}{The \link{DisableServiceRequest} object to pass to this method}

\item{serviceName}{Name of the service to disable}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/service.management
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/service.management)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-management/}{Google Documentation}

Other DisableServiceRequest functions: \code{\link{DisableServiceRequest}}
}
