% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_functions.R
\name{services.enable}
\alias{services.enable}
\title{Enable a managed service for a project with default setting.Operation<response: EnableServiceResponse>google.rpc.Status errors may contain agoogle.rpc.PreconditionFailure error detail.}
\usage{
services.enable(EnableServiceRequest, serviceName)
}
\arguments{
\item{EnableServiceRequest}{The \link{EnableServiceRequest} object to pass to this method}

\item{serviceName}{Name of the service to enable}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/service.management
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/service.management)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-management/}{Google Documentation}

Other EnableServiceRequest functions: \code{\link{EnableServiceRequest}}
}
