% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_functions.R
\name{services.generateConfigReport}
\alias{services.generateConfigReport}
\title{Generates and returns a report (errors, warnings and changes fromexisting configurations) associated withGenerateConfigReportRequest.new_valueIf GenerateConfigReportRequest.old_value is specified,GenerateConfigReportRequest will contain a single ChangeReport based on thecomparison between GenerateConfigReportRequest.new_value andGenerateConfigReportRequest.old_value.If GenerateConfigReportRequest.old_value is not specified, this methodwill compare GenerateConfigReportRequest.new_value with the last pushedservice configuration.}
\usage{
services.generateConfigReport(GenerateConfigReportRequest)
}
\arguments{
\item{GenerateConfigReportRequest}{The \link{GenerateConfigReportRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/service.management
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/service.management)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-management/}{Google Documentation}

Other GenerateConfigReportRequest functions: \code{\link{GenerateConfigReportRequest.newConfig}},
  \code{\link{GenerateConfigReportRequest.oldConfig}},
  \code{\link{GenerateConfigReportRequest}}
}
