% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_functions.R
\name{services.getIamPolicy}
\alias{services.getIamPolicy}
\title{Gets the access control policy for a resource.Returns an empty policy if the resource exists and does not have a policyset.}
\usage{
services.getIamPolicy(GetIamPolicyRequest, resource)
}
\arguments{
\item{GetIamPolicyRequest}{The \link{GetIamPolicyRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which the policy is being requested}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/service.management
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/service.management)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-management/}{Google Documentation}

Other GetIamPolicyRequest functions: \code{\link{GetIamPolicyRequest}}
}
