% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_functions.R
\name{services.list}
\alias{services.list}
\title{Lists managed services.Returns all public services. For authenticated users, also returns allservices the calling user has 'servicemanagement.services.get' permissionfor.**BETA:** If the caller specifies the `consumer_id`, it returns only theservices enabled on the consumer. The `consumer_id` must have the formatof 'project:{PROJECT-ID}'.}
\usage{
services.list(producerProjectId = NULL, consumerId = NULL,
  pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{producerProjectId}{Include services produced by the specified project}

\item{consumerId}{Include services consumed by the specified consumer}

\item{pageToken}{Token identifying which result to start with; returned by a previous list}

\item{pageSize}{Requested size of the next page of data}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/service.management
\item https://www.googleapis.com/auth/service.management.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only, https://www.googleapis.com/auth/service.management, https://www.googleapis.com/auth/service.management.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/service-management/}{Google Documentation}
}
