% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{SetIamPolicyRequest}
\alias{SetIamPolicyRequest}
\title{SetIamPolicyRequest Object}
\usage{
SetIamPolicyRequest(updateMask = NULL, policy = NULL)
}
\arguments{
\item{updateMask}{OPTIONAL: A FieldMask specifying which fields of the policy to modify}

\item{policy}{REQUIRED: The complete policy to be applied to the `resource`}
}
\value{
SetIamPolicyRequest object
}
\description{
SetIamPolicyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for `SetIamPolicy` method.
}
\seealso{
Other SetIamPolicyRequest functions: \code{\link{services.setIamPolicy}}
}
