% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Step}
\alias{Step}
\title{Step Object}
\usage{
Step(description = NULL, status = NULL)
}
\arguments{
\item{description}{The short description of the step}

\item{status}{The status code}
}
\value{
Step object
}
\description{
Step Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the status of one operation step.
}
