% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{SubmitConfigSourceRequest}
\alias{SubmitConfigSourceRequest}
\title{SubmitConfigSourceRequest Object}
\usage{
SubmitConfigSourceRequest(validateOnly = NULL, configSource = NULL)
}
\arguments{
\item{validateOnly}{Optional}

\item{configSource}{The source configuration for the service}
}
\value{
SubmitConfigSourceRequest object
}
\description{
SubmitConfigSourceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for SubmitConfigSource method.
}
