% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{SubmitConfigSourceResponse}
\alias{SubmitConfigSourceResponse}
\title{SubmitConfigSourceResponse Object}
\usage{
SubmitConfigSourceResponse(serviceConfig = NULL)
}
\arguments{
\item{serviceConfig}{The generated service configuration}
}
\value{
SubmitConfigSourceResponse object
}
\description{
SubmitConfigSourceResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for SubmitConfigSource method.
}
