% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{TrafficPercentStrategy.percentages}
\alias{TrafficPercentStrategy.percentages}
\title{TrafficPercentStrategy.percentages Object}
\usage{
TrafficPercentStrategy.percentages()
}
\value{
TrafficPercentStrategy.percentages object
}
\description{
TrafficPercentStrategy.percentages Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Maps service configuration IDs to their corresponding traffic percentage.Key is the service configuration ID, Value is the traffic percentagewhich must be greater than 0.0 and the sum must equal to 100.0.
}
\seealso{
Other TrafficPercentStrategy functions: \code{\link{TrafficPercentStrategy}}
}
