% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{TrafficPercentStrategy}
\alias{TrafficPercentStrategy}
\title{TrafficPercentStrategy Object}
\usage{
TrafficPercentStrategy(TrafficPercentStrategy.percentages = NULL,
  percentages = NULL)
}
\arguments{
\item{TrafficPercentStrategy.percentages}{The \link{TrafficPercentStrategy.percentages} object or list of objects}

\item{percentages}{Maps service configuration IDs to their corresponding traffic percentage}
}
\value{
TrafficPercentStrategy object
}
\description{
TrafficPercentStrategy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Strategy that specifies how Google Service Control should selectdifferentversions of service configurations based on traffic percentage.One example of how to gradually rollout a new service configuration usingthisstrategy:Day 1    Rollout {      id: 'example.googleapis.com/rollout_20160206'      traffic_percent_strategy {        percentages: {          'example.googleapis.com/20160201': 70.00          'example.googleapis.com/20160206': 30.00        }      }    }Day 2    Rollout {      id: 'example.googleapis.com/rollout_20160207'      traffic_percent_strategy: {        percentages: {          'example.googleapis.com/20160206': 100.00        }      }    }
}
\seealso{
Other TrafficPercentStrategy functions: \code{\link{TrafficPercentStrategy.percentages}}
}
