% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Type}
\alias{Type}
\title{Type Object}
\usage{
Type(options = NULL, fields = NULL, name = NULL, oneofs = NULL,
  sourceContext = NULL, syntax = NULL)
}
\arguments{
\item{options}{The protocol buffer options}

\item{fields}{The list of fields}

\item{name}{The fully qualified message name}

\item{oneofs}{The list of types appearing in `oneof` definitions in this type}

\item{sourceContext}{The source context}

\item{syntax}{The source syntax}
}
\value{
Type object
}
\description{
Type Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A protocol buffer message type.
}
