% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{UndeleteServiceResponse}
\alias{UndeleteServiceResponse}
\title{UndeleteServiceResponse Object}
\usage{
UndeleteServiceResponse(service = NULL)
}
\arguments{
\item{service}{Revived service resource}
}
\value{
UndeleteServiceResponse object
}
\description{
UndeleteServiceResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for UndeleteService method.
}
