% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{Usage}
\alias{Usage}
\title{Usage Object}
\usage{
Usage(requirements = NULL, producerNotificationChannel = NULL,
  rules = NULL)
}
\arguments{
\item{requirements}{Requirements that must be satisfied before a consumer project can use the}

\item{producerNotificationChannel}{The full resource name of a channel used for sending notifications to the}

\item{rules}{A list of usage rules that apply to individual API methods}
}
\value{
Usage object
}
\description{
Usage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration controlling usage of a service.
}
