% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicemanagement_objects.R
\name{VisibilityRule}
\alias{VisibilityRule}
\title{VisibilityRule Object}
\usage{
VisibilityRule(restriction = NULL, selector = NULL)
}
\arguments{
\item{restriction}{A comma-separated list of visibility labels that apply to the `selector`}

\item{selector}{Selects methods, messages, fields, enums, etc}
}
\value{
VisibilityRule object
}
\description{
VisibilityRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A visibility rule provides visibility configuration for an individual APIelement.
}
